公司业务相关/财务/requirements.txt
pandas
openpyxl
python-docx
pymupdf
pytesseract
Pillow
fastapi
uvicorn
python-multipart
jinja2
sqlmodel
requests
# Auth dependencies (Phase 5B)
passlib[bcrypt]
python-jose[cryptography]
email-validator
